+++ src/stubs/module/src/SettingsListener.php
@@ -5,6 +5,11 @@
 use YOOtheme\Config;
 use YOOtheme\Path;
 
+/**
+ * Settings listener.
+ *
+ * @see https://yootheme.com/support/yootheme-pro/joomla/developers-modules#add-settings
+ */
 class SettingsListener
 {
     public static function initCustomizer(Config $config)