+++ src/stubs/single-element/templates/content.php
@@ -1,11 +1,11 @@
 <?php if ($props['title'] || $props['content']): ?>
 <div>
 
-    <?php if ($props['title']): ?>
+    <?php if ($props['title'] != ''): ?>
     <h1><?= $props['title'] ?></h1>
     <?php endif; ?>
 
-    <?php if ($props['content']): ?>
+    <?php if ($props['content'] != ''): ?>
     <div><?= $props['content'] ?></div>
     <?php endif; ?>